refactor: split app_events.py god-handler into sibling handler mixins
Behavior-preserving structural decomposition of the residual ~980 LOC
god-handler in the app_frame handlers package (issue #535). The independent
feature clusters become behavior-named sibling mixins composed into AppFrame
exactly like the existing handlers, typed by the shared AppFrameProto:

- research.py -> DeckResearchHandlers (format/archetype selection, deck
  filters, archetype/deck load callbacks, summary, _load_decks)
- deck_content.py -> DeckContentHandlers (deck selection, copy/save/load,
  content render, daily-average build + callbacks)
- child_windows.py -> ChildWindowHandlers (tracker/timer/history/metagame/
  top-cards/radar windows, feedback dialog, window-change/close)
- data_loading.py -> DataLoadingHandlers (builder search, ensure_card_data,
  collection/bulk-index callbacks, background radar)

Stateless formatting statics move into a module-level deck_formatting.py
helper, and the ~15 controller state-delegation property/setter pairs move
into AppFramePropertiesMixin. app_events.py is removed; handlers/__init__.py
and AppFrame's composition are updated. No logic changes.

Adds tests/test_deck_formatting.py covering the extracted pure helpers.
